[WIP] Let Introduction be optional
Fixes #14

Just to track what I'm working on, I tried changing the initial route to be either the intro or the first page of the form, but there seems to be some initialization that hasn't been done at the point the router is invoked. Or something. I'll keep working on it.